(min invasive) sampling settings are part of the function.
This solves many problems and simplifies things. For reasons of
architecture, it might be attractive, to separate sampling from the
network of functions.
However, with the current foundation on exprtk this is not possible in a
consistent manner.

Consider this:
The sampling settings don't only influence one specific function, but
also every function that reference it in their formula...
For reasons of efficiency and simplicity the collection of functions
must be compiled into a static network, where the references between
functions are hardwired by exprtk.
After an expression is parsed via exprtk, there is no way to change the
resulting AST or inject information. The only way is to parse again,
which is very inefficient and leads to a pile of other problems.